refactor: 统一排名积分数据类型为Integer
- 将ActivityRankVO中的rank和score字段类型改为Integer
- 将ActivityStatsVO中的所有排名和积分字段类型改为Integer
- 将ActivityRankDTO中的rank和score字段类型改为Integer
- 重构ActivityServiceImpl中getUserRankScore方法返回Integer数组
- 更新相关使用方法中的类型转换逻辑，避免不必要的类型转换
- 修改AtomicLong为AtomicInteger，保持类型一致性
- 与数据库int unsigned类型保持一致，提升性能和内存使用效率

+++ bytelogs-api/src/main/java/top/harrylei/forum/api/model/rank/dto/ActivityRankDTO.java
@@ -30,10 +30,10 @@ public class ActivityRankDTO {
     /**
      * 排名
      */
-    private Long rank;
+    private Integer rank;
 
     /**
      * 积分
      */
-    private Double score;
+    private Integer score;
 }
